+++ CHANGELOG.md
@@ -171,9 +171,10 @@
 - Added a slorm-reaper list page
 
 ## Release 0.1.5 TBD
+### Added
+- Title is now dinamically updated
 
 ## TODO
- - changer de titre en fonction de la page build
  - ajouter page par défaut
  - fixes
  - mise à jour fichiers du jeu
@@ -186,5 +187,4 @@
 - copy reaper does not work on firefox
 
 ## Planned changes
-- slorm-reapers / slorm-legendaries ?
-- copy item as image ?
+- slorm-legendaries ?